feat: main and tasks window
