+++ cmd/tt/ui/utils/utils.go
