--- pkg/ui/ui.go
@@ -1,1 +1,0 @@
-package ui
